+++ model_navigator/framework_api/runners/runner_manager.py
@@ -94,7 +94,9 @@ class RunnerManager(DataObject):
             from model_navigator.framework_api.runners.trt import TrtRunner
 
             if runtime == RuntimeProvider.TRT_EXEC:
-                return TrtexecRunner(model_path, Format.TENSORRT, runtime_config=TrtExecRuntimeConfig(tf32=True))
+                return TrtexecRunner(
+                    model_path, Format.TENSORRT, runtime_config=TrtExecRuntimeConfig(tf32=True, use_cuda_graph=True)
+                )
             else:
                 return TrtRunner(EngineFromBytes(BytesFromPath(model_path)))
         elif format in (Format.TORCHSCRIPT, Format.TORCH_TRT):